backend/infra/cdk.out/asset.409b7eee35246093a9e68b83c49ae34b9edf848e31b1c51f33eee6dee5c239e9/requirements.txt
pyjwt==2.6.0
typing-extensions==3.7.4.1
